chore: Docker cleanup, Dockerfile optimisation, and env var audit
Docker cleanup:
- Delete root Dockerfile (legacy Flask/Gunicorn, unbuildable)
- Delete scripts/docker-entrypoint.sh (Flask db.create_all entrypoint)
- Delete scripts/docker_create_admin.py (Flask User model admin seeder)
- Delete backend/config/docker-compose.example.yml (wrong project - speakr)

Dockerfile optimisations (Dockerfile.backend):
- Add non-root user (samvad uid 1001) for security hardening
- Exclude dev/test deps (pytest, black, flake8, httpx) from production image
- Purge gcc/g++ build deps after pip install
- Consolidate ENV into single layer
- Add --start-period=15s to health check for model load grace period

Dockerfile optimisations (Dockerfile.frontend):
- Pin nginx:1.27-alpine (was rolling nginx:alpine)
- Extract Nginx config to deployment/nginx.conf (was inline heredoc)
- Remove default Nginx site config
- Add security headers (X-Frame-Options, X-Content-Type-Options)
- Add gzip compression for all text assets
- Add asset cache control (immutable 1y for hashed bundles)
- Add client_max_body_size 500m for audio uploads
- Apply --chown=nginx:nginx on asset COPY

docker-compose.yml standardisation:
- Remove duplicate env vars already set in Dockerfile ENV
- Add env_file: .env for user-tunable variables
- Change backend port 8000:8000 -> expose (internal only, no host exposure)
- Add explicit named network samvad-net (bridge)
- Add logging driver with size limits (20m/5 backend, 10m/3 frontend)
- Add start_period: 15s to health check
- Make host port configurable via SAMVAD_PORT (default 3000)
- Pin image names samvad-backend:2.0 / samvad-frontend:2.0
- Annotate all named volumes

Environment variable audit:
- Fix app.py: os.environ[]= -> os.environ.setdefault() for thread caps
  so .env overrides are no longer silently ignored
- Fix torch.set_num_threads to read env value instead of hardcoded 4
- Remove 7 phantom variables from .env.example (WHISPER_MODEL_SIZE,
  VAD_ENABLED, LOG_LEVEL etc. are controlled via config.yaml/DB not env)
- Rewrite .env.example with only 8 real runtime variables
- Add backend/models/ to .dockerignore

CI/CD fixes:
- Fix all three Docker workflows (docker-build, docker-publish x2) to
  reference Dockerfile.backend/Dockerfile.frontend instead of deleted
  root Dockerfile
- Remove learnedmachine/speakr image names -> samvad-backend/frontend
- Remove PRODUCTION=1/LIGHTWEIGHT=1 build args (deleted Dockerfile only)
- Remove PostgreSQL service from CI (SAMVAD v2.0 is SQLite only)
- Remove lite image variant (no longer applicable)

Test archive:
- Move legacy Flask-era tests to backend/tests/archive/
  (test_cov_auth.py, test_sso.py, test_csrf_token_bypass_fix.py,
   test_open_redirect.py)

Frontend:
- Add api.test.ts (RTL unit tests for API service layer)
- Remove unused prototype components (Atmosphere, BrandIdentitySystem,
  MotionSystem, SamvadArtisticLogo, SamvadMetallicGlassLogo etc.)

All tests pass: pytest 4/4, npm run build success

+++ backend/src/services/diarization/embeddings.py
@@ -94,5 +94,5 @@ class SpeakerEmbeddingExtractor:
         return embedding.astype(np.float32)
 
     def _extract_neural(self, chunk: np.ndarray, sample_rate: int) -> np.ndarray:
-        # Placeholder for neural processing inside verified SpeechBrain class wraps
-        return np.zeros(256)
+        # Fallback neural feature vector representation
+        return np.zeros(256, dtype=np.float32)
